Move install-related tests from snapstate_test.go to snapstate_install_test.go
